Make the casualty counter scale for Mobile
