+++ src/pages/gamemode/components/GamemodeToolbar.tsx
@@ -1,0 +1,195 @@
+import { ChevronLeft, ChevronRight } from "@mui/icons-material";
+import { Button, ButtonGroup, Stack } from "@mui/material";
+import Box from "@mui/material/Box";
+import Toolbar from "@mui/material/Toolbar";
+import Typography from "@mui/material/Typography";
+import { FaSkullCrossbones } from "react-icons/fa";
+import { GiCrackedShield } from "react-icons/gi";
+import { useNavigate } from "react-router-dom";
+import { armyListData } from "../../../assets/data.ts";
+import { SquareIconButton } from "../../../components/common/icon-button/SquareIconButton.tsx";
+import { useRosterInformation } from "../../../hooks/useRosterInformation.ts";
+import { useScreenSize } from "../../../hooks/useScreenSize.ts";
+import { useGameModeState } from "../../../state/gamemode";
+
+export const GamemodeToolbar = () => {
+  const navigate = useNavigate();
+  const screen = useScreenSize();
+  const { gameState, updateGameState, endGame } = useGameModeState();
+  const { roster, getAdjustedMetaData } = useRosterInformation();
+
+  const game = gameState[roster.id];
+
+  const updateCasualties = (update: 1 | -1) => {
+    updateGameState(roster.id, {
+      casualties: game.casualties + update,
+    });
+  };
+
+  const armyListMetadata = armyListData[roster.armyList];
+  const metadata = getAdjustedMetaData(roster);
+  const breakPoint =
+    metadata.units -
+    Math.floor(metadata.units * (1 - (armyListMetadata.break_point ?? 0.49)));
+  const quarter = metadata.units - Math.floor(metadata.units * 0.25);
+  const casualties = game.casualties + game.heroCasualties;
+
+  return screen.isMobile ? (
+    <>
+      <Stack direction="row" justifyContent="space-between" sx={{ my: 1 }}>
+        <Button
+          variant="outlined"
+          color="inherit"
+          startIcon={<ChevronLeft />}
+          onClick={() => {
+            if (window.history.state && window.history.state.idx > 0) {
+              navigate(-1);
+            } else {
+              navigate("/gamemode/start", { replace: true }); // the current entry in the history stack will be replaced with the new one with { replace: true }
+            }
+          }}
+        >
+          Back
+        </Button>
+        <Button
+          variant="outlined"
+          onClick={() => {
+            endGame(roster.id);
+            navigate("/gamemode/start");
+          }}
+        >
+          End game
+        </Button>
+      </Stack>
+      <Stack direction="row" justifyContent="space-around" sx={{ m: 2 }}>
+        <Typography color={breakPoint - casualties <= 0 ? "error" : "inherit"}>
+          Until Broken:{" "}
+          <b>
+            {breakPoint - casualties > 0 ? (
+              breakPoint - casualties
+            ) : (
+              <GiCrackedShield />
+            )}
+          </b>
+        </Typography>
+        <Typography color={quarter - casualties <= 0 ? "error" : "inherit"}>
+          Until Quartered:{" "}
+          <b>
+            {quarter - casualties > 0 ? (
+              quarter - casualties
+            ) : (
+              <FaSkullCrossbones />
+            )}
+          </b>
+        </Typography>
+      </Stack>
+      <Stack direction="row" gap={2} justifyContent="center">
+        <Typography variant="h6" className="middle-earth">
+          Casualties:
+        </Typography>
+        <SquareIconButton
+          onClick={() => updateCasualties(-1)}
+          icon={<ChevronLeft />}
+          iconColor="white"
+          backgroundColor="darkgrey"
+          iconPadding=".3rem"
+          disabled={game.casualties === 0}
+        />
+        <Typography variant="h6" className="middle-earth" sx={{ mx: 1 }}>
+          {casualties}
+        </Typography>
+        <SquareIconButton
+          onClick={() => updateCasualties(+1)}
+          icon={<ChevronRight />}
+          iconColor="white"
+          backgroundColor="darkgrey"
+          iconPadding=".3rem"
+        />
+      </Stack>
+    </>
+  ) : (
+    <Toolbar
+      sx={{
+        backgroundColor: "rgba(211,211,211,0.5)",
+        mb: 2,
+        justifyContent: "space-between",
+      }}
+    >
+      <Box>
+        <ButtonGroup>
+          <Button
+            startIcon={<ChevronLeft />}
+            onClick={() => {
+              if (window.history.state && window.history.state.idx > 0) {
+                navigate(-1);
+              } else {
+                navigate("/gamemode/start", { replace: true }); // the current entry in the history stack will be replaced with the new one with { replace: true }
+              }
+            }}
+          >
+            Back
+          </Button>
+          <Button
+            onClick={() => {
+              endGame(roster.id);
+              navigate("/gamemode/start");
+            }}
+          >
+            End game
+          </Button>
+        </ButtonGroup>
+      </Box>
+      <Stack direction="row" gap={2}>
+        <Typography variant="h6" className="middle-earth">
+          Casualties:
+        </Typography>
+        <SquareIconButton
+          onClick={() => updateCasualties(-1)}
+          icon={<ChevronLeft />}
+          iconColor="white"
+          backgroundColor="darkgrey"
+          iconPadding=".3rem"
+          disabled={game.casualties === 0}
+        />
+        <Typography variant="h6" className="middle-earth" sx={{ mx: 1 }}>
+          {casualties}
+        </Typography>
+        <SquareIconButton
+          onClick={() => updateCasualties(+1)}
+          icon={<ChevronRight />}
+          iconColor="white"
+          backgroundColor="darkgrey"
+          iconPadding=".3rem"
+        />
+      </Stack>
+      <Box>
+        <Typography
+          sx={{ display: "flex", justifyContent: "space-between", gap: 2 }}
+          color={breakPoint - casualties <= 0 ? "error" : "inherit"}
+        >
+          Until Broken:{" "}
+          <b>
+            {breakPoint - casualties > 0 ? (
+              breakPoint - casualties
+            ) : (
+              <GiCrackedShield />
+            )}
+          </b>
+        </Typography>
+        <Typography
+          sx={{ display: "flex", justifyContent: "space-between", gap: 2 }}
+          color={quarter - casualties <= 0 ? "error" : "inherit"}
+        >
+          Until Quartered:{" "}
+          <b>
+            {quarter - casualties > 0 ? (
+              quarter - casualties
+            ) : (
+              <FaSkullCrossbones />
+            )}
+          </b>
+        </Typography>
+      </Box>
+    </Toolbar>
+  );
+};
